+++ backend/task.hpp
@@ -1,82 +1,34 @@
-#if !defined(TASK_)
-#define TASK_
-
-#include <string>
+#if !defined(TASK)
+#define TASK
 
 #include "nlohmann/json.hpp"
+#include <string>
 
 using nlohmann::json;
 using std::string;
 
-namespace task {
-    enum ScriptType { PYTHON,
-                      BASH,
-                      UNINIT };
+enum TaskScriptType { PYTHON, BASH, UNINIT };
 
-    enum Status { IDLE,
-                  RUNNING,
-                  UNINIT };
+enum TaskStatus { IDLE, RUNNING, UNINIT };
 
-    constexpr char *TASK_KEY_ID = "id";
-    constexpr char *TASK_KEY_TITLE = "title";
-    constexpr char *TASK_KEY_SCRIPTCODE = "scriptCode";
-    constexpr char *TASK_KEY_SCRIPTTYPE = "scriptType";
-    constexpr char *TASK_KEY_STATUS = "status";
-    constexpr char *TASK_KEY_PID = "pid";
-    constexpr char *TASK_KEY_INTERVALINSEC = "intervalInSec";
-    constexpr char *TASK_KEY_MAXTIMES = "maxTimes";
-    constexpr char *TASK_KEY_EXITCODE = "exitCode";
-    constexpr char *TASK_KEY_EXITTIMESTAMP = "exitTimeStamp";
+struct Task {
+    int32_t id;
+    string title;
+    string scriptCode;
+    TaskScriptType scriptType;
+    TaskStatus status;
+    int32_t pid;
+    int32_t intervalInSec;
+    int32_t maxTimes;
+    int32_t timesExecuted;
+    int32_t exitCode;
+    int64_t exitTimeStamp;
 
-    struct Task {
-        int32_t id;
-        string title;
-        string scriptCode;
-        ScriptType scriptType;
-        Status status;
-        int32_t pid;
-        int32_t intervalInSec;
-        int32_t maxTimes;
-        int32_t timesExecuted;
-        int32_t exitCode;
-        int64_t exitTimeStamp;
+    Task() {}
 
-        Task() {}
+    Task(json data);
 
-        Task(json data) {
-            try {
-                id = data[TASK_KEY_ID].get<int32_t>();
-                title = data[TASK_KEY_TITLE].get<string>();
-                scriptCode = data[TASK_KEY_SCRIPTCODE].get<string>();
-                scriptType = data[TASK_KEY_SCRIPTTYPE].get<ScriptType>();
-                status = data[TASK_KEY_STATUS].get<Status>();
-                pid = data[TASK_KEY_PID].get<int32_t>();
-                intervalInSec = data[TASK_KEY_INTERVALINSEC].get<int32_t>();
-                maxTimes = data[TASK_KEY_MAXTIMES].get<int32_t>();
-                exitCode = data[TASK_KEY_EXITCODE].get<int32_t>();
-                exitTimeStamp = data[TASK_KEY_EXITTIMESTAMP].get<int64_t>();
-            } catch (json::exception) {
-                throw std::runtime_error("create task from json failed");
-            }
-        }
+    json toJson() const;
+};
 
-        json toJson() const {
-            json res = {
-                {TASK_KEY_ID, id},
-                {TASK_KEY_TITLE, title},
-                {TASK_KEY_SCRIPTCODE, scriptCode},
-                {TASK_KEY_SCRIPTTYPE, scriptType},
-                {TASK_KEY_STATUS, status},
-                {TASK_KEY_PID, pid},
-                {TASK_KEY_INTERVALINSEC, intervalInSec},
-                {TASK_KEY_MAXTIMES, maxTimes},
-                {TASK_KEY_EXITCODE, exitCode},
-                {TASK_KEY_EXITTIMESTAMP, exitTimeStamp},
-            };
-            return res;
-        }
-    };
-
-} // namespace task
-
-#endif // TASK_
+#endif // TASK
