--- backend/task_db_mgr.cpp
@@ -1,89 +1,0 @@
-
-#include "task_db_mgr.hpp"
-
-namespace task_db_mgr {
-
-    using std::string, std::max, std::to_string;
-    using task::Task;
-
-    void DBGuard::init() {
-        DBGuard guard;
-        guard.open(true);
-        guard.writeToDB(DB_KEY_TASK_ID_SET, "[]");
-    }
-
-    bool checkInDB(DBGuard &guard, int32_t taskId) {
-        auto idListJson = json::parse(guard.readDB(DB_KEY_TASK_ID_SET));
-        bool found = false;
-        for (size_t i = 0; i < idListJson.size(); i++) {
-            if (idListJson[i].get<int32_t>() == taskId) {
-                idListJson.erase(i);
-                found = true;
-                break;
-            }
-        }
-        return found;
-    }
-
-    void createTask(const string &title, const string &scriptCode, int32_t scriptType,
-                    int32_t intervalInSec = 0, int32_t maxTimes = 1) {
-        Task newTask;
-        newTask.title = title;
-        newTask.scriptCode = scriptCode;
-        newTask.scriptType = static_cast<task::ScriptType>(scriptType);
-        newTask.status = task::IDLE;
-        newTask.pid = 0;
-        newTask.intervalInSec = intervalInSec;
-        newTask.maxTimes = maxTimes;
-        newTask.timesExecuted = 0;
-        newTask.exitCode = 0;
-        newTask.exitTimeStamp = 0;
-
-        DBGuard guard;
-        guard.open();
-        auto idListJson = json::parse(guard.readDB(DB_KEY_TASK_ID_SET));
-        int32_t maxId = 0;
-        for (auto &&v : idListJson) {
-            maxId = max(maxId, v.get<int32_t>());
-        }
-        newTask.id = maxId + 1;
-        idListJson.push_back(newTask.id);
-        guard.writeToDB(DB_KEY_TASK_ID_SET, idListJson.dump());
-        guard.writeToDB(to_string(newTask.id), newTask.toJson().dump());
-    }
-
-    void deleteTask(int32_t taskId) {
-        DBGuard guard;
-        guard.open();
-        auto idListJson = json::parse(guard.readDB(DB_KEY_TASK_ID_SET));
-        bool found = false;
-        for (size_t i = 0; i < idListJson.size(); i++) {
-            if (idListJson[i].get<int32_t>() == taskId) {
-                idListJson.erase(i);
-                found = true;
-                break;
-            }
-        }
-        if (!found)
-            throw std::runtime_error("db delete failed: task " + to_string(taskId) + " not exist");
-        guard.writeToDB(DB_KEY_TASK_ID_SET, idListJson.dump());
-    }
-
-    void updateTask(int32_t taskId, const Task &newTask) {
-        DBGuard guard;
-        guard.open();
-        auto idListJson = json::parse(guard.readDB(DB_KEY_TASK_ID_SET));
-        bool found = false;
-        for (size_t i = 0; i < idListJson.size(); i++) {
-            if (idListJson[i].get<int32_t>() == taskId) {
-                idListJson.erase(i);
-                found = true;
-                break;
-            }
-        }
-        if (!found)
-            throw std::runtime_error("db delete failed: task " + to_string(taskId) + " not exist");
-        guard.writeToDB(to_string(taskId), newTask.toJson().dump());
-    }
-
-} // namespace task_db_mgr
